# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "NISQ Analyzer API",
    "contact" : {
      "name" : "GitHub Repository",
      "url" : "https://github.com/UST-QuAntiL/nisq-analyzer"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.4.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5010/nisq-analyzer",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/execution-results/{resultId}" : {
      "get" : {
        "tags" : [ "execution-result" ],
        "description" : "Retrieve a single execution result",
        "operationId" : "getExecutionResult",
        "parameters" : [ {
          "name" : "resultId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecutionResultDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/qpu-selection-results/jobs/{resId}" : {
      "get" : {
        "tags" : [ "qpu-selection-result" ],
        "description" : "Retrieve a single QPU selection job",
        "operationId" : "getQpuSelectionJob",
        "parameters" : [ {
          "name" : "resId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QpuSelectionJobDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/qpu-selection-results/{resId}" : {
      "get" : {
        "tags" : [ "qpu-selection-result" ],
        "description" : "Retrieve a single QPU selection result",
        "operationId" : "getQpuSelectionResult",
        "parameters" : [ {
          "name" : "resId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QpuSelectionResultDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/qpu-selection-results/" : {
      "get" : {
        "tags" : [ "qpu-selection-result" ],
        "description" : "Retrieve all QPU selection results",
        "operationId" : "getQpuSelectionResults",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QpuSelectionResultListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/qpu-selection-results/jobs" : {
      "get" : {
        "tags" : [ "qpu-selection-result" ],
        "description" : "Retrieve all QPU selection jobs",
        "operationId" : "getQpuSelectionJobs",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QpuSelectionJobListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/qpu-selection-results/{resId}/execute" : {
      "post" : {
        "tags" : [ "qpu-selection-result" ],
        "description" : "Execute a compilation result",
        "operationId" : "executeQpuSelectionResult",
        "parameters" : [ {
          "name" : "resId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExecuteAnalysisResultRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecutionResultDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/selection" : {
      "post" : {
        "tags" : [ "root" ],
        "description" : "Select implementations for an algorithm",
        "operationId" : "selectImplementations",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectionRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisJobDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/qpu-selection" : {
      "post" : {
        "tags" : [ "root" ],
        "description" : "Select the most suitable quantum computer for a quantum circuit loaded from the given URL",
        "operationId" : "selectQpuForCircuitFile_1",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/qpuselection_body"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/QpuSelectionDto"
              }
            },
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/QpuSelectionDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QpuSelectionJobDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AnalysisResultDto" : {
        "type" : "object",
        "properties" : {
          "implementation" : {
            "$ref" : "#/components/schemas/ImplementationDto"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "inputParameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "originalCircuitResultId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "qpuSelectionJobId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "correlationId" : {
            "type" : "string"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "Links" : {
        "type" : "object",
        "additionalProperties" : {
          "$ref" : "#/components/schemas/Link"
        }
      },
...
      "AnalysisJobDto" : {
        "type" : "object",
        "properties" : {
          "analysisResultList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AnalysisResultDto"
            }
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "implementedAlgorithm" : {
            "type" : "string",
            "format" : "uuid"
          },
          "time" : {
            "type" : "string",
            "format" : "date-time"
          },
          "inputParameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "ready" : {
            "type" : "boolean"
          },
          "initialMcdaJob" : {
            "type" : "string",
            "format" : "uuid"
          },
          "initialMcdaMethod" : {
            "type" : "string"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "ExecutionResultDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "INITIALIZED", "RUNNING", "FAILED", "FINISHED" ]
          },
          "statusCode" : {
            "type" : "string"
          },
          "result" : {
            "type" : "string"
          },
          "shots" : {
            "type" : "integer",
            "format" : "int32"
          },
          "histogramIntersectionValue" : {
            "type" : "number",
            "format" : "double"
          },
          "resultLocation" : {
            "type" : "string",
            "format" : "uri"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "QpuSelectionJobDto" : {
        "type" : "object",
        "properties" : {
          "qpuSelectionResultList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QpuSelectionResultDto"
            }
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "time" : {
            "type" : "string",
            "format" : "date-time"
          },
          "circuitName" : {
            "type" : "string"
          },
          "ready" : {
            "type" : "boolean"
          },
          "userId" : {
            "type" : "string"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "QpuSelectionResultDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "provider" : {
            "type" : "string"
          },
          "qpu" : {
            "type" : "string"
          },
          "compiler" : {
            "type" : "string"
          },
          "analyzedDepth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "analyzedWidth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "analyzedTotalNumberOfOperations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "analyzedNumberOfSingleQubitGates" : {
            "type" : "integer",
            "format" : "int32"
          },
          "analyzedNumberOfMultiQubitGates" : {
            "type" : "integer",
            "format" : "int32"
          },
          "analyzedNumberOfMeasurementOperations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "analyzedMultiQubitGateDepth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "numberOfQubits" : {
            "type" : "integer",
            "format" : "int32"
          },
          "t1" : {
            "type" : "number",
            "format" : "float"
          },
          "t2" : {
            "type" : "number",
            "format" : "float"
          },
          "avgReadoutError" : {
            "type" : "number",
            "format" : "float"
          },
          "avgSingleQubitGateError" : {
            "type" : "number",
            "format" : "float"
          },
          "avgMultiQubitGateError" : {
            "type" : "number",
            "format" : "float"
          },
          "avgSingleQubitGateTime" : {
            "type" : "number",
            "format" : "float"
          },
          "avgMultiQubitGateTime" : {
            "type" : "number",
            "format" : "float"
          },
          "maxGateTime" : {
            "type" : "number",
            "format" : "float"
          },
          "simulator" : {
            "type" : "boolean"
          },
          "time" : {
            "type" : "string",
            "format" : "date-time"
          },
          "queueSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "transpiledCircuit" : {
            "type" : "string"
          },
          "transpiledLanguage" : {
            "type" : "string"
          },
          "circuitName" : {
            "type" : "string"
          },
          "qpuSelectionJobId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string"
          },
          "originalCircuitResultId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "estimatedHistogramIntersectionValue" : {
            "type" : "number",
            "format" : "float"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "QpuSelectionResultListDto" : {
        "type" : "object",
        "properties" : {
          "qpuSelectionResultList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QpuSelectionResultDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "QpuSelectionJobListDto" : {
        "type" : "object",
        "properties" : {
          "qpuSelectionJobList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QpuSelectionJobDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "ExecuteAnalysisResultRequestDto" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string"
          },
          "tokens" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          },
          "correlationId" : {
            "type" : "string"
          }
        }
      },
...
      "SelectionRequestDto" : {
        "type" : "object",
        "properties" : {
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "allowedProviders" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "compilers" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "preciseResultsPreference" : {
            "type" : "boolean"
          },
          "shortWaitingTimesPreference" : {
            "type" : "boolean"
          },
          "queueImportanceRatio" : {
            "type" : "number",
            "format" : "float"
          },
          "maxNumberOfCompiledCircuits" : {
            "type" : "integer",
            "format" : "int32"
          },
          "predictionAlgorithm" : {
            "type" : "string"
          },
          "metaOptimizer" : {
            "type" : "string"
          },
          "tokens" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          },
          "algorithmId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "mcdaMethodName" : {
            "type" : "string"
          },
          "mcdaWeightLearningMethod" : {
            "type" : "string"
          }
        }
      },
...
      "QpuSelectionDto" : {
        "type" : "object",
        "properties" : {
          "allowedProviders" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "circuitLanguage" : {
            "type" : "string"
          },
          "circuitUrl" : {
            "type" : "string",
            "format" : "url"
          },
          "qasmCode" : {
            "type" : "string"
          },
          "tokens" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          },
          "refreshToken" : {
            "type" : "string"
          },
          "circuitName" : {
            "type" : "string"
          },
          "preciseResultsPreference" : {
            "type" : "boolean"
          },
          "shortWaitingTimesPreference" : {
            "type" : "boolean"
          },
          "queueImportanceRatio" : {
            "type" : "number",
            "format" : "float"
          },
          "maxNumberOfCompiledCircuits" : {
            "type" : "integer",
            "format" : "int32"
          },
          "predictionAlgorithm" : {
            "type" : "string"
          },
          "metaOptimizer" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "compilers" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "qpuselection_body" : {
        "type" : "object",
        "properties" : {
          "qpuSelectionDto" : {
            "$ref" : "#/components/schemas/QpuSelectionDto"
          },
          "circuit" : {
            "type" : "string",
            "format" : "binary"
          }
        }
      },
...
    }
  }
}